Add Sass `{outputStyle:'expanded'}` for cmsScss gulp task.
For CloudCannon to Accept a custom stylesheet, the stylesheet needs to have fully expanded code.
```css
p.myClass {
  color: #000;
}
```
The gulp task cmsScss needs the output CSS to be fully expanded and not compressed.
To change the output css to expanded, we need to pass a settings object to sass() with `outputStyle: expanded`.